Important changes! Re-semantic the TreeList.
This is typically dangerous thing to do. Since these sortable is
quite fragile. But this enables to move items more smoothly. And Fix
the dummy index 0 thing which was problematic since we can't move
items above index 1, it stuck due to using <details> tag as sortable
container, and inevitably had <summary> at index 0 at a dummy list item.
This commit fixes that.