Fix new module schema is not object type.
The problem is a brand new module cannot initiate any new model since
it was not an object type.

This commit makes sure it is every time. The fix could be to make model
schema as object as module initialization (we will do that), but this
fix does not require re-creating module, thus, a non-breaking fix.