Basic ability to encode jsonschema into module
This commit is just PoC renderable, without breaking. There might be
other samples that make it break. But here we test with aws-cfn;
https://github.com/aws-cloudformation/aws-cfn-lint-visual-studio-code/blob/master/schema/all-spec.json
spec that is almost 3 MB. Final rendered DOM is about 9 MB!

It's renderable, looks fine on server, but such lots of DOM nodes make
it unusable on browser (e.g. sortablejs can't handle it, dom patch is
slow)

In next commits, we will paginate models and figure out UX.