Introduce ProfileLive and break the MainLive
Currently MainLive is rendering /:file_id which is the whole single
schema. Since we just introduce projects table that split the single
schema into one main and list of model schemas. We need a new way to
render these schemas. We no longer render the giantic schema with
MainLive. This commit literally break the MainLive page.