Encode fragment ref as anchor.
This should be most-of-the-time case for import. User's schema,
most of the time, define $ref as "#/$definition/<def>". We automatically
resolve it as model anchor.

For schema that has $ref points to non-1st level $defs/$definition,
We are not suporting it. We will put note for user to break their depth
definition into flat, 1st level of $defs.