Return {pre,post,new_map} for Sch.put
Previously it only returns new_map which is normally a whole schema
tree, and most of the time it's quite big.

Also returning "post" which is an updated container sch for a given path
has advantage of small payload to be broadcasted.

It's also not convincing to introduce any event sourcing here.