Fix massive models update on Main#move
It should only update the relevant models, not all. The cost is that
Main#broadcast_update_sch requires two db queries, though, we don't
care batch update for now since it's rare that ones need to move from/to
a bunch of models at the same time. Ideally, it sohuld be just 1 or 2
models involved.